added discount data to the order aggregate

+++ src/Services/Coupon/Coupon.Application/Features/Coupon/Models/CouponDto.cs
@@ -1,3 +1,3 @@
 namespace Coupon.Application.Features.Coupon.Models;
 
-public record CouponDto(Guid Id, string Code, int Discount);
+public record CouponDto(Guid Id, string Code, decimal Discount);
