+++ src/Services/Coupon/Coupon.Domain/AggregatesModel/Coupon.cs
@@ -6,11 +6,11 @@ public class Coupon : BaseEntity
 {
     public Guid Id { get; set; }
     public string Code { get; private set; }
-    public int Discount { get; private set; }
+    public decimal Discount { get; private set; }
     public bool Consumed { get; private set; }
     public int? OrderId { get; private set; }
 
-    public Coupon(string code, int discount)
+    public Coupon(string code, decimal discount)
     {
         Id = Guid.NewGuid();
         Code = code;
